+++ python/lsst/ts/hvac/csc.py
@@ -27,7 +27,6 @@ import re
 
 import numpy as np
 
-from .config_schema import CONFIG_SCHEMA
 from . import __version__
 from .enums import (
     CommandItem,
@@ -136,14 +135,12 @@ class InternalItemState:
         return recent_values
 
 
-class HvacCsc(salobj.ConfigurableCsc):
+class HvacCsc(salobj.BaseCsc):
     """Commandable SAL Component for the HVAC (Heating, Ventilation and Air
     Conditioning).
 
     Parameters
     ----------
-    config_dir : `string`
-        The configuration directory
     initial_state : `salobj.State`
         The initial state of the CSC
     simulation_mode : `int`
@@ -168,23 +165,18 @@ class HvacCsc(salobj.ConfigurableCsc):
 
     def __init__(
         self,
-        config_dir=None,
         initial_state=salobj.State.STANDBY,
         simulation_mode=0,
         settings_to_apply="",
         start_telemetry_publishing=True,
     ):
-        self.config = None
-        self._config_dir = config_dir
         self._add_config_commands()
         super().__init__(
             name="HVAC",
             index=0,
-            config_schema=CONFIG_SCHEMA,
-            config_dir=config_dir,
             initial_state=initial_state,
-            simulation_mode=simulation_mode,
             settings_to_apply=settings_to_apply,
+            simulation_mode=simulation_mode,
         )
 
         self.mqtt_client = None
@@ -200,6 +192,10 @@ class HvacCsc(salobj.ConfigurableCsc):
         # and this gets initialized in the connect method.
         self.hvac_state = None
 
+        # The host and port to connect to.
+        self.host = ""
+        self.port = 0
+
         # Helper for reading the HVAC data
         self.xml = MqttInfoReader()
 
@@ -213,10 +209,7 @@ class HvacCsc(salobj.ConfigurableCsc):
         simulation mode.
         """
         self.log.info("Connecting.")
-        self.log.info(self.config)
         self.log.info(f"self.simulation_mode = {self.simulation_mode}")
-        if self.config is None:
-            raise RuntimeError("Not yet configured")
         # if self.connected:
         #     raise RuntimeError("Already connected")
 
@@ -229,8 +222,10 @@ class HvacCsc(salobj.ConfigurableCsc):
             self.mqtt_client = SimClient(self.start_telemetry_publishing)
         else:
             # Use the MQTT Client.
-            self.log.info(f"Connecting MqttClient to {self.config.host}.")
-            self.mqtt_client = MqttClient(host=self.config.host, port=self.config.port)
+            self.log.info(
+                f"Connecting MqttClient  to host {self.host} and port {self.port}."
+            )
+            self.mqtt_client = MqttClient(host=self.host, port=self.port)
 
         await self.mqtt_client.connect()
         if self.start_telemetry_publishing:
@@ -313,9 +308,6 @@ class HvacCsc(salobj.ConfigurableCsc):
         if self.connected:
             await self.disconnect()
         await super().end_disable(id_data)
-
-    async def configure(self, config):
-        self.config = config
 
     def _get_topic_enabled_state(self, topic):
         """Determine whether a device represented by the MQTT topic is enabled