+++ tests/test_hvac_csc.py
@@ -48,11 +48,10 @@ flake8.configure_logging(1)
 
 class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     def basic_make_csc(
-        self, initial_state, config_dir, simulation_mode, settings_to_apply, **kwargs
+        self, initial_state, simulation_mode, settings_to_apply, **kwargs
     ):
         return hvac.HvacCsc(
             initial_state=initial_state,
-            config_dir=config_dir,
             simulation_mode=simulation_mode,
             settings_to_apply=settings_to_apply,
             start_telemetry_publishing=False,
@@ -61,7 +60,6 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     async def test_standard_state_transitions(self):
         async with self.make_csc(
             initial_state=salobj.State.STANDBY,
-            config_dir=None,
             simulation_mode=1,
             settings_to_apply="",
         ):
@@ -81,7 +79,6 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     async def test_version(self):
         async with self.make_csc(
             initial_state=salobj.State.STANDBY,
-            config_dir=None,
             simulation_mode=1,
             settings_to_apply="",
         ):
@@ -136,7 +133,6 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     async def test_enable_on_all_subsystems_one_by_one(self):
         async with self.make_csc(
             initial_state=salobj.State.STANDBY,
-            config_dir=None,
             simulation_mode=1,
             settings_to_apply="",
         ):
@@ -228,7 +224,6 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     async def test_config(self):
         async with self.make_csc(
             initial_state=salobj.State.STANDBY,
-            config_dir=None,
             simulation_mode=1,
             settings_to_apply="",
         ):