+++ doc/version_history.rst
@@ -11,6 +11,7 @@ v0.6.0
 
 * Replaced the use of ts_salobj functions with ts_utils functions.
 * Added auto-enable capability.
+* Converted the CSC to a non-configurable CSC.
 
 Requires: